requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Flask==3.1.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
proglog==0.1.10
requests==2.32.3
setuptools==75.5.0
tqdm==4.67.0
urllib3==2.2.3
Werkzeug==3.1.3
yt-dlp==2024.11.4
